+++ src/main/java/soccerfriend/exception/exception/DuplicatedException.java
@@ -1,14 +1,14 @@
-package soccerfriend.exception.member;
+package soccerfriend.exception.exception;
 
 import lombok.Getter;
 import soccerfriend.exception.ExceptionCode;
 
 @Getter
-public class NotExistException extends RuntimeException {
+public class DuplicatedException extends RuntimeException {
 
     private ExceptionCode exceptionCode;
 
-    public NotExistException(ExceptionCode exceptionCode) {
+    public DuplicatedException(ExceptionCode exceptionCode) {
         this.exceptionCode = exceptionCode;
     }
 }
