+++ src/main/java/soccerfriend/exception/exception/NotMatchException.java
@@ -1,14 +1,14 @@
-package soccerfriend.exception.member;
+package soccerfriend.exception.exception;
 
 import lombok.Getter;
 import soccerfriend.exception.ExceptionCode;
 
 @Getter
-public class DuplicatedException extends RuntimeException {
+public class NotMatchException extends IllegalArgumentException {
 
     private ExceptionCode exceptionCode;
 
-    public DuplicatedException(ExceptionCode exceptionCode) {
+    public NotMatchException(ExceptionCode exceptionCode) {
         this.exceptionCode = exceptionCode;
     }
 }
