+++ src/main/java/soccerfriend/service/MemberService.java
@@ -4,10 +4,9 @@ import lombok.RequiredArgsConstructor;
 import org.mindrot.jbcrypt.BCrypt;
 import org.springframework.stereotype.Service;
 import soccerfriend.dto.Member;
-import soccerfriend.exception.member.DuplicatedException;
-import soccerfriend.exception.member.NotMatchException;
+import soccerfriend.exception.exception.DuplicatedException;
+import soccerfriend.exception.exception.NotMatchException;
 import soccerfriend.mapper.MemberMapper;
-import utility.InputForm;
 import utility.InputForm.UpdatePasswordRequest;
 
 import java.util.Optional;
